requirements.txt
Flask==2.3.3
pycryptodome==3.19.0
protobuf==3.20.3
requests==2.31.0
google==3.0.0